Putting back df in session_state at app.py, created operations.py
