+++ data_analysis/operations.py
@@ -1,5 +1,19 @@
 import pandas as pd
-from .stats_functions import ft_std, ft_percentile
+import math
+
+def ft_std(series: pd.Series, mean: float, count_val: int) -> float:
+    """Calculate the standard deviation of the series"""
+    summand = 0
+    for val in series:
+        if pd.notna(val):  # ignore NaN / None
+            summand += (val - mean)**2
+
+    return math.sqrt(summand / count_val)
+
+
+def ft_percentile(p: float, count: int) -> float:
+    """Calculate the p-th percentile of the series"""
+    return p * (count + 1)
 
 def operations(series: pd.Series) -> list:
     count_val  = 0
@@ -35,29 +49,3 @@ def operations(series: pd.Series) -> list:
     
     return [count_val, mean_val, std_val,
             min_val, p25_val, p50_val, p75_val, max_val]
-
-def stats_table(df: pd.DataFrame, whitelist=None) -> pd.DataFrame:
-    """
-    df: pandas DataFrame to analyze
-    whitelist: list of column names to include. If None, use all columns.
-    """
-    # Determine which columns to use
-    if whitelist is None:
-        columns_to_use = df.columns
-    else:
-        # Keep only columns that exist in df
-        columns_to_use = [col for col in whitelist if col in df.columns]
-
-    rows = []
-
-    for col in columns_to_use:
-        series = df[col]
-        rows.append(operations(series))
-
-    stats = pd.DataFrame(
-        rows,
-        index=columns_to_use,
-        columns=['count','mean','std','min','P25','P50','P75','max']
-    ).T
-
-    return stats
