+++ pages/Describe.py
@@ -3,10 +3,10 @@ from data_analysis.data_analysis import DataAnalysis
 
 def show_describe():
     """Display descriptive statistics from the uploaded CSV."""
+
     if "data_analysis" in st.session_state:
         data_analysis = st.session_state.data_analysis
-
-        if data_analysis.csv_uploaded:
+        if data_analysis.df is not None:
             described_data = data_analysis.get_stats()
             st.dataframe(described_data)
         else:
@@ -14,5 +14,4 @@ def show_describe():
     else:
         st.write("Please upload a CSV file in the main app to see the data description.")
 
-# Describe function
 show_describe()
